+++ packages/mikro-orm/CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## next (YYYY-MM-DD)
 
+* Refactor: rename `MikroOperationBobbin` to `MikroResolverFactory`
+
 ## 0.5.0 (2024-12-03)
 
 * Chore: update `@standard-schema/spec` to 1.0.0-beta.4